+++ src/components/filter_group/filter_group.tsx
@@ -18,7 +18,7 @@ export type EuiFilterGroupProps = HTMLAttributes<HTMLDivElement> &
      */
     fullWidth?: boolean;
     /**
-     *  When `true` creates a shorter height filter group
+     *  When `true` creates a shorter height filter group matching that of `compressed` form controls
      */
     compressed?: boolean;
   };